Move status bar and progress dialogs into own classes
